Move Fleeting Luck scene control button into fleeting-luck.js.
